Added logout functionality and temporary isRegistered
Added buttons with logout functionality on homescreen stack. It calls
authCtx.logout when pressed. Also added a temporary workaround for
checking if user is registered by setting it on device when registration
done. For development purposes for registration details page  we can remove the line that actually sets the AsyncToken as True. In future we will need to call a backend API after login success, to check if the user is registered or not. Then take directly to home screen.